requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.1
django-filter==23.5
requests==2.31.0
python-dotenv==1.0.0
drf-yasg==1.21.7